Add dummy route with https
This commit adds custom certificates for 127.0.0.1 and localhost which
are to be used in conjunction with the Docker setup. Running
'docker-compose up' will let you run 'curl https://127.0.0.1/' and the
Clojure server will respond with a dummy JSON object. The whole thing is
powered by an nginx reverse proxy generated by the other docker compose
service.